# api-gateway/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Cinema Booking API
  description: A microservice-based cinema ticket booking system
  version: 1.0.0
  contact:
    name: Cinema Booking System
    email: support@cinema-booking.com
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/auth/verify:
    post:
      summary: Verify JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  valid:
                    type: boolean
                    example: true
        '401':
          description: Invalid token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        role:
          type: string
          example: customer
...